types: clear every truncating pointer declaration (128 -> 0)
The PC port's crashes have all been one defect class: a declaration that says
s32 or u32 where the real thing is a pointer. Host pointers are eight bytes
and host stacks live above 4 GiB, so the low half sign-extends and the first
dereference faults. tools/pc/lp64_audit.py in HALBERD lists them -- Makefile.pc
builds with -w, which is what hid the class -- and the truncating tier is now
empty.

Reconciled against their definitions, roughly 120 call sites across 25 files.
The largest were func_801E3A84_ovl13 and func_801E19D0_ovl15, declared
(s32,s32,s32) at 66 sites while every caller passes D_800DFBD0[objId][n],
which track_arrays.h declares struct DObj **[]. Also func_800B6144,
func_800AA608, func_800A4F48, func_800A8100, func_80117904, func_800F88C8,
func_801A3E80_ovl7, the enelib func_8019B*/func_80199F1C family, and
func_800B2340 -- whose fix required deleting 14 now-wrong (s32)(uintptr_t)
casts from its guarded arms, since a corrected prototype would otherwise
sign-extend the handle they produce.

TWO THINGS THAT WERE NOT FIXED, DELIBERATELY:

func_800A22D4 keeps its s32 spelling in ovl7_4.c. Its definition takes
UnkA22A8 * and dereferences it, and every argument passed is a real pointer,
but retyping the declaration moves 14 words of ovl7_4 -- IDO stages a pointer
argument straight into $a0 while the ROM builds it in $a1 and copies. The
game was compiled against the narrower prototype. Reverted, call sites
truncate explicitly, measurement recorded at the declaration.

The Shape28 joint word stays a truncated pointer. Widening unk8/unk30/unk58
in the ovl13/ovl15 anim-control structs is a PORT REGRESSION, not a fix:
those are views over ovl2_9.c's Shape28 arena, which ovl2_8.c requires to
stay 40 bytes and offset-stable on LP64, and func_8010E740 rebuilds the
pointer by zero-extension. Field widths reverted, stores made explicit, and
all four files that carry the struct now say so.

Effect on the port: the smoke test went from SIGSEGV at 19.16s in
func_8019F410_ovl7 to attract-loop-complete -- gGameState route
0>1>2>3>4>5>6>7>8>9>2>3>4>5>6 -- alive at the 60s timeout, 8 of 8 runs. The
lane could not attribute that to a single declaration by bisection and did
not claim one.

ROM sha1 6cea2d46b929 unchanged.

+++ src/ovl2/plylib.c
@@ -32,6 +32,11 @@ struct UnkPos4C {
 };
 
 GObj *func_800A8234(s32, s32, s32);
+/* src/ovl1/ovl1.c:5409 defines this as func_800A22D4(UnkA22A8 *). The u32
+ * spelling is kept here because this TU's other three call sites pass
+ * gKirbyState.unk4C/unk50/unk6C, which are u32 handle fields; the two
+ * pointer sites truncate explicitly (lossless -- the -no-pie image keeps
+ * the whole arena below 4 GiB). */
 void func_800A22D4(u32);
 /* defined in ovl1_10.c; was re-declared four times in this file as
  * (u32, s32), which disagrees with the definition and left the call in
@@ -4056,7 +4061,7 @@ void func_80121A04(void) {
         } else {
             var_a0 = D_8012EAE0;
             if (var_a0 != NULL) {
-                func_800A22D4(var_a0);
+                func_800A22D4((u32) (uintptr_t) var_a0);
                 D_8012EAE0 = NULL;
                 var_a0 = NULL;
             }
@@ -4069,7 +4074,7 @@ void func_80121A04(void) {
     } else {
         D_8012EADC = 0;
         if (D_8012EAE0 != NULL) {
-            func_800A22D4(D_8012EAE0);
+            func_800A22D4((u32) (uintptr_t) D_8012EAE0);
             D_8012EAE0 = NULL;
         }
     }
